pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rice-sdk"
version = "0.1.0"
description = "Python SDK for Rice"
authors = [{ name = "Rice Team" }]
requires-python = ">=3.8"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-mock>=3.12.0", "black", "isort", "mypy"]
